Moved class stuff to class.rs
 - Added example and documenation
