+++ boa/examples/classes.rs
@@ -1,16 +1,11 @@
 use boa::{
-    builtins::{
-        object::{Class, ClassBuilder},
-        property::Attribute,
-        value::Value,
-    },
+    builtins::{property::Attribute, value::Value},
+    class::{Class, ClassBuilder},
     exec::Interpreter,
     forward_val,
     realm::Realm,
-    Result,
+    Finalize, Result, Trace,
 };
-
-use gc::{Finalize, Trace};
 
 // We create a new struct that is going to represent a person.
 //
@@ -140,7 +135,7 @@ fn main() {
 		if (!Person.is('Hello')) {
 			console.log('\'Hello\' string is not a Person class instance.');
 		}
-        
+
         console.log(Person.staticProperty);
         console.log(person.inheritedProperty);
 	    console.log(Person.prototype.inheritedProperty === person.inheritedProperty);
