+++ boa/src/lib.rs
@@ -35,6 +35,7 @@
 )]
 
 pub mod builtins;
+pub mod class;
 pub mod environment;
 pub mod exec;
 pub mod profiler;
